requirements.txt
APScheduler==2.1.0
Django==1.4.3
Fabric==1.6.0
Jinja2==2.6
PyYAML==3.10
Pygments==1.6
South==0.7.6
Sphinx==1.1.3
beautifulsoup4==4.1.3
coverage==3.6
cssselect==0.8
distribute==0.6.36
django-admin-bootstrapped==0.3.2
django-appconf==0.6
django-auth-ldap==1.1.4
django-bootstrap-toolkit==2.8.0
django-compress==1.0.1
django-compressor==1.2
django-crispy-forms==1.2.3
django-dajax==0.9.2
django-dajaxice==0.5.5
django-debug-toolbar==0.9.4
django-dynamic-fixture==1.6.4
django-extensions==1.1.1
django-gravatar==0.1.0
django-tastypie==0.9.12
docutils==0.10
git-remote-helpers==0.1.0
howdoi==1.1.2
httpretty==0.5.12
ipython==0.13.2
lxml==3.1.0
mechanize==0.2.5
mimeparse==0.1.3
paramiko==1.10.1
pexpect==2.4
py-bcrypt==0.2
pyasn1==0.1.6
pycrypto==2.6
pygraphviz==1.1
pyobjc==2.5.1
pyobjc-core==2.5.1
pyobjc-framework-Accounts==2.5.1
pyobjc-framework-AddressBook==2.5.1
pyobjc-framework-AppleScriptKit==2.5.1
pyobjc-framework-AppleScriptObjC==2.5.1
pyobjc-framework-Automator==2.5.1
pyobjc-framework-CFNetwork==2.5.1
pyobjc-framework-CalendarStore==2.5.1
pyobjc-framework-Cocoa==2.5.1
pyobjc-framework-Collaboration==2.5.1
pyobjc-framework-CoreData==2.5.1
pyobjc-framework-CoreLocation==2.5.1
pyobjc-framework-CoreText==2.5.1
pyobjc-framework-DictionaryServices==2.5.1
pyobjc-framework-EventKit==2.5.1
pyobjc-framework-ExceptionHandling==2.5.1
pyobjc-framework-FSEvents==2.5.1
pyobjc-framework-InputMethodKit==2.5.1
pyobjc-framework-InstallerPlugins==2.5.1
pyobjc-framework-InstantMessage==2.5.1
pyobjc-framework-LatentSemanticMapping==2.5.1
pyobjc-framework-LaunchServices==2.5.1
pyobjc-framework-Message==2.5.1
pyobjc-framework-PreferencePanes==2.5.1
pyobjc-framework-PubSub==2.5.1
pyobjc-framework-QTKit==2.5.1
pyobjc-framework-Quartz==2.5.1
pyobjc-framework-ScreenSaver==2.5.1
pyobjc-framework-ScriptingBridge==2.5.1
pyobjc-framework-SearchKit==2.5.1
pyobjc-framework-ServerNotification==2.5.1
pyobjc-framework-ServiceManagement==2.5.1
pyobjc-framework-Social==2.5.1
pyobjc-framework-SyncServices==2.5.1
pyobjc-framework-SystemConfiguration==2.5.1
pyobjc-framework-WebKit==2.5.1
pyquery==1.2.4
pysnmp==4.2.4
python-dateutil==1.5
python-ldap==2.4.10
pytz==2012j
pyzmq==13.0.2
requests==1.1.0
six==1.2.0
slumber==0.5.3
sphinx-http-domain==0.2
webkit2png==0.8.2
wsgiref==0.1.2
